Move BrickProvider from com.wayfair.brickkit.brick to com.wayfair.brickkit.viewholder.factory
Fixes https://github.com/wayfair/brickkit-android/issues/181
